Improved query optimizer
