+++ graphene/contrib/django/tests/test_optimizequery.py
@@ -1,41 +1,16 @@
-from functools import wraps
-from graphql.core.utils.get_field_def import get_field_def
 import pytest
 
 import graphene
 from graphene.contrib.django import DjangoObjectType
 
-from graphql.core.type.definition import GraphQLList, GraphQLNonNull
-
-
 from ..tests.models import Reporter
 from ..debug.plugin import DjangoDebugPlugin
+from ..fetcher import fetch_only_required, get_fields
 
 # from examples.starwars_django.models import Character
 
 pytestmark = pytest.mark.django_db
 
-def get_fields(info):
-    field_asts = info.field_asts[0].selection_set.selections
-    only_args = []
-    _type = info.return_type
-    if isinstance(_type, (GraphQLList, GraphQLNonNull)):
-        _type = _type.of_type
-
-    for field in field_asts:
-        field_def = get_field_def(info.schema, _type, field)
-        f = field_def.resolver
-        fetch_field = getattr(f, 'django_fetch_field', None)
-        if fetch_field:
-            only_args.append(fetch_field)
-    return only_args
-
-def fetch_only_required(f):
-    @wraps(f)
-    def wrapper(*args):
-        info = args[-1]
-        return f(*args).only(*get_fields(info))
-    return wrapper
 
 def test_should_query_well():
     r1 = Reporter(last_name='ABA')
